requirements.txt
scapy
paramiko
requests
colorama
dnspython
python-nmap
